Factor out postPriorityEntriesHandler
